# src/OpenMetrc.Scraper/OpenApi/metrc-swagger-v2.yaml
swagger: 2.0
info:
  title: METRC API
  version: v2
paths:
...
  /employees/v2/:
    get:
      tags:
      - Employees
      summary: ''
      description: 'Retrieves a list of employees for a specified Facility.</br><b>Permissions Required</b>: Manage Employees • View Employees'
      externalDocs:
        x-doc-key: get_employees_v2_
      operationId: GetEmployees
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility under which to get the employees.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /employees/v2/permissions:
    get:
      tags:
      - Employees
      summary: ''
      description: 'Retrieves the permissions of a specified Employee, identified by their Employee License Number, for a given Facility.</br><b>Permissions Required</b>: Manage Employees'
      externalDocs:
        x-doc-key: get_employees_v2_permissions
      operationId: GetEmployeePermissions
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility under which to get the Employee.
        format: ''
      - type: string
        name: employeeLicenseNumber
        in: query
        required: true
        description: The License Number of the Employee under which to get the Permission.
        format: ''
      responses:
        200:
          description: Ok
...
  /items/v2/brand/{id}:
    delete:
      tags:
      - Items
      summary: ''
      description: 'Archives the specified Item Brand by Id for the given Facility License Number.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: delete_items_v2_brand_{id}
      operationId: DeleteItemBrandById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for the Item Brand to delete.
        format: ''
      responses:
        200:
          description: Ok
  /items/v2/{id}:
    delete:
      tags:
      - Items
      summary: ''
      description: 'Archives the specified Product by Id for the given Facility License Number.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: delete_items_v2_{id}
      operationId: DeleteItemById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for the Item to delete.
        format: ''
      responses:
        200:
          description: Ok
    get:
      tags:
      - Items
      summary: ''
      description: 'Retrieves detailed information about a specific Item by Id.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: get_items_v2_{id}
      operationId: GetItemById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Item will be validated against the specified License Number. If not specified, the Item will be validated against all of the User's current Facilities. Please note that if the Item is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /items/v2/active:
    get:
      tags:
      - Items
      summary: ''
      description: 'Returns a list of active items for the specified Facility.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: get_items_v2_active
      operationId: GetItemActive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of active items.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /items/v2/brands:
    get:
      tags:
      - Items
      summary: ''
      description: 'Retrieves a list of active item brands for the specified Facility.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: get_items_v2_brands
      operationId: GetItemBrands
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of active item brands.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /items/v2/categories:
    get:
      tags:
      - Items
      summary: ''
      description: 'Retrieves a list of item categories.</br><b>Permissions Required</b>: <i>none</i>'
      externalDocs:
        x-doc-key: get_items_v2_categories
      operationId: GetItemCategories
      parameters:
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the categories will be retrieved for the specified License Number. If not specified, all item categories will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /items/v2/file/{id}:
    get:
      tags:
      - Items
      summary: ''
      description: 'Retrieves a file by its Id for the specified Facility.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: get_items_v2_file_{id}
      operationId: GetItemFileById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: The License Number of the Facility for which to return the file.
        format: ''
      responses:
        200:
          description: Ok
  /items/v2/inactive:
    get:
      tags:
      - Items
      summary: ''
      description: 'Retrieves a list of inactive items for the specified Facility.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: get_items_v2_inactive
      operationId: GetItemInactive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of inactive items.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /items/v2/photo/{id}:
    get:
      tags:
      - Items
      summary: ''
      description: 'Retrieves an image by its Id for the specified Facility.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: get_items_v2_photo_{id}
      operationId: GetItemPhotoById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: The License Number of the Facility for which to return the image.
        format: ''
      responses:
        200:
          description: Ok
  /items/v2/:
    post:
      tags:
      - Items
      summary: ''
      description: 'Creates one or more new products for the specified Facility. NOTE: To include a photo with an item, first use POST /items/v2/photo to POST the photo, and then use the returned Id in the request body in this endpoint.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: post_items_v2_
      operationId: PostItems
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to create new items.
        format: ''
      responses:
        200:
          description: Ok
    put:
      tags:
      - Items
      summary: ''
      description: 'Updates one or more existing products for the specified Facility.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: put_items_v2_
      operationId: PutItems
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for the Item updates.
        format: ''
      responses:
        200:
          description: Ok
  /items/v2/brand:
    post:
      tags:
      - Items
      summary: ''
      description: 'Creates one or more new item brands for the specified Facility identified by the License Number.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: post_items_v2_brand
      operationId: PostItemBrand
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for the Item Brands to create.
        format: ''
      responses:
        200:
          description: Ok
    put:
      tags:
      - Items
      summary: ''
      description: 'Updates one or more existing item brands for the specified Facility.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: put_items_v2_brand
      operationId: PutItemBrand
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for the Item Brand updates.
        format: ''
      responses:
        200:
          description: Ok
  /items/v2/file:
    post:
      tags:
      - Items
      summary: ''
      description: 'Uploads one or more image or PDF files for products, labels, packaging, or documents at the specified Facility.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: post_items_v2_file
      operationId: PostItemFile
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: ''
        format: ''
      responses:
        200:
          description: Ok
  /items/v2/photo:
    post:
      tags:
      - Items
      summary: ''
      description: 'This endpoint allows only BMP, GIF, JPG, and PNG files and uploaded files can be no more than 5 MB in size.</br><b>Permissions Required</b>: Manage Items'
      externalDocs:
        x-doc-key: post_items_v2_photo
      operationId: PostItemPhoto
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to add an image.
        format: ''
      responses:
        200:
          description: Ok
...